Create mineos.py
I did not make this script, I'm not claiming that it is mine all i'm doing is maintaining it for the use on the Raspberry Pi.
The Dev's of MineOs (http://minecraft.codeemo.com/) deserve all the credit for this, I worked by William Dizon's guidance to track down the issue as to why this wouldn't work on Arm. Thank You William!